# 00_docs/openapi.yaml
openapi: 3.0.0
info:
  title: Flashcard API
  version: 1.0.0
  description: フラッシュカードアプリケーションのAPI仕様
servers:
  - url: https://api.example.com
    description: 本番環境
  - url: http://localhost:8787
    description: 開発環境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      description: 統一エラーレスポンス形式
      properties:
        code:
          type: string
          description: エラーを一意に識別するコード
          example: "INVALID_REQUEST"
          enum:
            - INVALID_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: "リクエストパラメータが不正です"
      required:
        - code
        - message
...
    CardImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
      required:
        - id
        - url
...
    CardResponse:
      type: object
      description: カードのAPIレスポンス形式
      properties:
        id:
          type: string
          format: uuid
        frontContent:
          type: string
        backContent:
          type: string
        frontImages:
          type: array
          items:
            $ref: "#/components/schemas/CardImage"
          description: カード表面の画像一覧
        backImages:
          type: array
          items:
            $ref: "#/components/schemas/CardImage"
          description: カード裏面の画像一覧
        tags:
          type: array
          items:
            type: string
        easeFactor:
          type: number
          format: float
          minimum: 1.3
        interval:
          type: integer
          minimum: 0
        nextStudyDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - frontContent
        - backContent
        - frontImages
        - backImages
        - tags
        - easeFactor
        - interval
        - nextStudyDate
        - createdAt
        - updatedAt
...
paths:
...
  /api/decks/{deckId}/study/cards:
    get:
      tags:
        - study
      summary: 学習対象カードの取得
      description: |
        指定されたデッキから学習すべきカードのリストを取得します。
        学習対象となるカードには以下のようなものが含まれます:
        1. 次回学習日が到来しているカード
        2. 未学習のカード
      security:
        - bearerAuth: []
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: 取得する学習カードの最大数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 学習カードの取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/CardResponse"
                required:
                  - cards
        "404":
          description: デッキが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...